Change to 64 bit default integer
When the basis gets too large 64 bit integers are needed for PARDISO,
and eventually the index pointers in sparse arrays. Currently this is
done with the -fdefault-integer-8 flag for gfortran. In the future it
would be better to use kind parameters for all integers, so that the
type is clear to the programmer and user. The program now requires
building stdlib and arpack with ILP64 support, and it links to the ILP64
version of MKL.

+++ src/tools/input_tools.f90
@@ -51,7 +51,7 @@ module input_tools
     real(dp) :: calc_start
     real(dp) :: calc_end
     real(dp) :: other_param
-    integer :: n_calc
+    integer(int32) :: n_calc
     integer :: n_quasi
     integer :: n_blocks(2)
